spring-cloud-playground-service/src/main/java/org/spring/cloud/playground/datetimeprovider/DateTimeProviderImpl.java
package org.spring.cloud.playground.datetimeprovider;

import org.springframework.stereotype.Component;

import java.time.LocalDate;
import java.time.LocalDateTime;
import java.time.LocalTime;

@Component
public class DateTimeProviderImpl implements DateTimeProvider {

    @Override
    public LocalDateTime getCurrentDateTime() {
        return LocalDateTime.now();
    }

    @Override
    public LocalDate getCurrentDate() {
        return LocalDate.now();
    }

    @Override
    public LocalTime getCurrentTime() {
        return LocalTime.now();
    }

}
